Add custom hooks and press enter key logic, fix throttle
